# perf: optimize decode bits fn