--- a/src/machine/asm/mod.rs
+++ b/src/machine/asm/mod.rs
@@ -191,6 +191,12 @@ impl Memory for Box<AsmCoreMachine> {
         self.load16(&(addr)).map(|v| v as u16)
     }
 
+    fn execute_load32(&mut self, addr: u64) -> Result<u32, Error> {
+        let page_indices = get_page_indices(addr, 4)?;
+        check_permission(self, &page_indices, FLAG_EXECUTABLE)?;
+        self.load32(&(addr)).map(|v| v as u32)
+    }
+
     fn load8(&mut self, addr: &u64) -> Result<u64, Error> {
         let addr = *addr;
         let page_indices = get_page_indices(addr, 1)?;