@quake 2021-03-10
> This has some boundary issues we need to deal with: assuming the last 2 bytes of an executable page is a C instruction, using the older way, such an instruction could be loaded and executed without problems, but the behavior would change after introducing this PR.

applied an address checking accordingly, the benchmark result is still valid, please review again, thanks.